fix little bug and refactor
